docs: rename folders
