+++ src/components/@common/heading-text/HeadingText.tsx
