+++ src/components/@common/heading-text/index.ts
